requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==3.2.18
djangorestframework==3.14.0
ecdsa==0.18.0
exceptiongroup==1.1.1
executing==1.2.0
fonttools==4.39.3
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wcwidth==0.2.6
webdriver-manager==3.8.5
wsproto==1.2.0

